requirements.txt
torch==2.2.0
torchvision==0.17.1
torchaudio==2.2.0
numpy==1.25.0
Pillow==10.0.0
matplotlib==3.8.0
opencv-python==4.9.0.76
scikit-learn==1.3.0
PyYAML==6.0
datasets==2.17.1
tqdm==4.66.0
